feat: enforce integer material quantities across all functions
Material quantities must be whole numbers everywhere — purchase (orderedQty),
issue (requestedQty), receive/transfer/goods-receive quantity, count (actualQty),
and adjust (quantityChange). Money fields (unitCost/standardCost) stay decimal.

- Backend: add zod .int() to every quantity field (approved/issuedQty were
  already int).
- Frontend: qty inputs use step="1"; client-component fields also block "."/","
  keys, matching the existing material-issue ItemsField pattern.
- Guard test quantityIntegerRule.test.ts asserts fractional is rejected and
  money stays fractional (integration tests bypass zod, so this locks the rule).
- Documented in business-rules.md §1.5 and CLAUDE.md.

Tests 42/42. api+web typecheck clean. Live-verified: fractional qty -> 400.

+++ apps/api/src/validation/goodsReceiveSchema.ts
@@ -8,7 +8,7 @@ export const createGoodsReceiveSchema = z.object({
     .array(
       z.object({
         materialId: z.string().min(1),
-        quantity: z.coerce.number().positive(),
+        quantity: z.coerce.number().int().positive(),
         unitCost: z.coerce.number().nonnegative(),
       }),
     )
